+++ src/user/user.controller.ts
@@ -1,6 +1,6 @@
 import { ID } from 'scalars';
 import { UserService } from './user.service';
-import { User } from '../shared/models/user/user.model';
+import { UserModel } from '../shared/models/user/user.model';
 import { CreateUserDto } from '../shared/models/user/create-user.dto';
 import { UpdateUserDto } from '../shared/models/user/update-user.dto';
 import { Controller, Get, Param, Post, Body, Put, Delete, HttpCode, HttpStatus } from '@nestjs/common';
@@ -10,17 +10,17 @@ export class UserController {
   constructor(private readonly userService: UserService) {}
 
   @Get(':id')
-  getUser(@Param('id') id: ID): User {
+  getUser(@Param('id') id: ID): UserModel {
     return this.userService.getUser(id);
   }
 
   @Post()
-  createUser(@Body() createUserDto: CreateUserDto): User {
+  createUser(@Body() createUserDto: CreateUserDto): UserModel {
     return this.userService.createUser(createUserDto);
   }
 
   @Put(':id')
-  updateUser(@Body() updateUserDto: UpdateUserDto, @Param('id') id: ID): User {
+  updateUser(@Body() updateUserDto: UpdateUserDto, @Param('id') id: ID): UserModel {
     return this.userService.updateUser(updateUserDto, id);
   }