+++ src/user/user.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { ID } from 'scalars';
-import { User } from '../shared/models/user/user.model';
 import { UserRepository } from './user.repository';
+import { UserModel } from '../shared/models/user/user.model';
 import { CreateUserDto } from '../shared/models/user/create-user.dto';
 import { UpdateUserDto } from '../shared/models/user/update-user.dto';
 
@@ -9,15 +9,15 @@ import { UpdateUserDto } from '../shared/models/user/update-user.dto';
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
 
-  public getUser(id: ID): User {
+  public getUser(id: ID): UserModel {
     return this.userRepository.getUser(id);
   }
 
-  public createUser(createUserDto: CreateUserDto): User {
+  public createUser(createUserDto: CreateUserDto): UserModel {
     return this.userRepository.addUser(createUserDto);
   }
 
-  public updateUser(updateUserDto: UpdateUserDto, id: ID): User {
+  public updateUser(updateUserDto: UpdateUserDto, id: ID): UserModel {
     return this.userRepository.updateUser(updateUserDto, id);
   }