+++ CHANGELOG.md
@@ -6,11 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `User` model class name to be `UserModel`.
+
 ## [0.3.0] - 2019-08-06
 
 ### Added
 
-- Global typings module
+- Global typings module.
 - Relation between `Post` and `User`.
 
 ## [0.2.0] - 2019-08-03